Fix searchbar focus
Fixes https://github.com/gentakojima/emojione-picker-ubuntu/issues/18

set_keep_above raises the window to the front but doesn't focus it yet
Since the window gets created at the location of the mouse, xdotool click should now focus it.
For some reason, the window only actually gets focus when calling get_position() first. This does not seem to be a timing issue as replacing get_position with a sleep does not work.